requirements.txt
Django==4.1.1
django-crum==0.7.9
django-widget-tweaks==1.4.12
et-xmlfile==1.0.1
flake8==5.0.4
pandas==1.5.1
Pillow==9.1.1
python-dotenv==0.19.0
XlsxWriter==1.1.5
xmltodict==0.13.0